RedditClone/backend/requirements.txt
marshmallow
PyJWT
alembic==1.0.10
aniso8601==7.0.0
Click==7.0
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==0.15.4
